+++ bin/cli.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+const mdLinks = require('../index.js');
+
+console.log('hi')
